# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Referral Demo API
  version: 2.0.0
servers:
  - url: http://localhost:3000/api
paths:
  /health:
    get:
      responses:
        '200': { description: Healthy }
  /auth/login:
    post:
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginInput' } } }
      responses:
        '200': { description: Authenticated, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user, content: { application/json: { schema: { $ref: '#/components/schemas/AuthUser' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/reset-password:
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ResetPasswordInput' } } }
      responses:
        '200': { description: Password reset and old token invalidated, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/me/invitation:
    post:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Existing invitation, content: &invitationContent { application/json: { schema: { $ref: '#/components/schemas/InvitationLink' } } } }
        '201': { description: Created invitation, content: *invitationContent }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /invitations/{token}:
    get:
      parameters: [{ $ref: '#/components/parameters/InvitationToken' }]
      responses:
        '200': { description: Public invitation, content: { application/json: { schema: { $ref: '#/components/schemas/PublicInvitation' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /invitations/{token}/accept:
    post:
      parameters: [{ $ref: '#/components/parameters/InvitationToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AcceptInvitationInput' } } }
      responses:
        '201': { description: Invitation accepted; temporaryPassword is returned once, content: { application/json: { schema: { $ref: '#/components/schemas/AcceptInvitationResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/me/referral-summary:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user's referral summary, content: { application/json: { schema: { $ref: '#/components/schemas/ReferralSummary' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    InvitationToken:
      name: token
      in: path
      required: true
      schema: { type: string, pattern: '^[A-HJ-NP-Z2-9]{12}$' }
  responses:
    BadRequest: &errorResponse { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Unauthorized: { <<: *errorResponse, description: Authentication required or invalid credentials }
    Forbidden: { <<: *errorResponse, description: Password reset required }
    NotFound: { <<: *errorResponse, description: Resource not found }
    Conflict: { <<: *errorResponse, description: Email already registered }
  schemas:
    LoginInput:
      type: object
      required: [email, password]
      properties: { email: { type: string, format: email, maxLength: 254 }, password: { type: string, minLength: 1, maxLength: 72, writeOnly: true } }
      additionalProperties: false
    ResetPasswordInput:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword: { type: string, maxLength: 72, writeOnly: true }
        newPassword: { type: string, minLength: 10, maxLength: 72, writeOnly: true }
        confirmPassword: { type: string, minLength: 10, maxLength: 72, writeOnly: true }
      additionalProperties: false
    AuthUser:
      type: object
      required: [id, name, email, mustResetPassword]
      properties: { id: { type: string }, name: { type: string }, email: { type: string, format: email }, mustResetPassword: { type: boolean } }
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties: { accessToken: { type: string }, user: { $ref: '#/components/schemas/AuthUser' } }
    InvitationLink:
      type: object
      required: [token, path, publicUrl]
      properties: { token: { type: string, pattern: '^[A-HJ-NP-Z2-9]{12}$' }, path: { type: string }, publicUrl: { type: string, format: uri } }
    PublicInvitation:
      type: object
      required: [inviter]
      properties:
        inviter:
          type: object
          required: [name]
          properties: { name: { type: string } }
          additionalProperties: false
      additionalProperties: false
    AcceptInvitationInput:
      type: object
      required: [name, email]
      properties: { name: { type: string, minLength: 1, maxLength: 80 }, email: { type: string, format: email, maxLength: 254 } }
      additionalProperties: false
    AcceptInvitationResult:
      type: object
      required: [user, referral, temporaryPassword]
      properties:
        user: { $ref: '#/components/schemas/User' }
        referral:
          type: object
          required: [id, inviterName, rewardCredits, acceptedAt]
          properties: { id: { type: string }, inviterName: { type: string }, rewardCredits: { type: integer, minimum: 1 }, acceptedAt: { type: string, format: date-time } }
        temporaryPassword: { type: string, writeOnly: true }
    User:
      type: object
      required: [id, name, email]
      properties: { id: { type: string }, name: { type: string }, email: { type: string, format: email } }
    ReferralSummary:
      type: object
      required: [id, name, email, creditBalance, successfulReferralCount, invitation, referrals, hasMore]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        creditBalance: { type: integer, minimum: 0 }
        successfulReferralCount: { type: integer, minimum: 0 }
        invitation:
          oneOf: [{ $ref: '#/components/schemas/SummaryInvitation' }, { type: 'null' }]
        referrals:
          type: array
          maxItems: 100
          items:
            type: object
            required: [id, inviteeName, rewardCredits, acceptedAt]
            properties: { id: { type: string }, inviteeName: { type: string }, rewardCredits: { type: integer, minimum: 1 }, acceptedAt: { type: string, format: date-time } }
        hasMore: { type: boolean }
    SummaryInvitation:
      type: object
      required: [token, publicUrl]
      properties: { token: { type: string }, publicUrl: { type: string, format: uri } }
    ApiError:
      type: object
      required: [statusCode, code, message, requestId, timestamp, path]
      properties:
        statusCode: { type: integer }
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
        path: { type: string }